+++ src/restx_monkey/__init__.py
@@ -1,4 +1,3 @@
-import pathlib
 import typing
 
 import pkg_resources
@@ -74,39 +73,5 @@ def patch_restx(
         flask_restx.reqparse.RequestParser = restx_reqparser.RequestParser
 
     if update_swagger_ui and is_incompatible:
-        import flask_restx
-        my_static = pathlib.Path(__file__).parent / "static"
-        my_static_files = my_static / "files"
-        lib_static = pathlib.Path(flask_restx.__file__).parent / "static"
-        lib_static_files = lib_static / "files"
-
-        # create missing
-        for folder in (lib_static, lib_static_files):
-            if not folder.exists():
-                try:
-                    folder.mkdir()
-                except PermissionError:
-                    pass
-        # replace old files in static
-        for my_path in my_static.iterdir():
-            if my_path.is_file():
-                try:
-                    my_path.replace(lib_static/my_path.name)
-                except PermissionError:
-                    pass
-
-        # replace old font files with new
-        for my_path in my_static_files.iterdir():
-            if my_path.is_file():
-                try:
-                    my_path.replace(lib_static_files/my_path.name)
-                except PermissionError:
-                    pass
-
-
-def main():
-    patch_restx()
-
-
-if __name__ == "__main__":
-    main()
+        from . import swagger_ui
+        swagger_ui.replace_static_swagger_files()