+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Patch to override `static` directory of `flask-restx` with last Swagger UI.
+
 ## [0.2.1] - 2022-08-31
 
 ### Modify